Fix checkstyle errors and add necessary JavaDocs
There are several classes that failed the checkstyle tests. Some
classes are missing Javadocs.

Fixing all those checkstyle errors and adding JavaDocs comments
improves the code quality.

As a step towards writing clean readable code, let's keep in mind
standard Java and Git practices when writing code. Doing so will
make the subsequent fixes easier.
