+++ build.gradle
@@ -50,7 +50,7 @@ test {
 application {
     // mainClass.set("seedu.duke.Duke")
     // mainClass.set("Launcher")
-    mainClass.set("Nat.Launcher")
+    mainClass.set("nat.Launcher")
 }
 
 shadowJar {
