+++ docs/README.md
@@ -1,4 +1,4 @@
-# Nat.Duke User Guide
+# nat.Duke User Guide
 
 // Update the title above to match the actual product name
