+++ src/Test/java/nat/ToDoTest.java
@@ -1,8 +1,8 @@
-package Nat;  //same package as the class being tested
-
-import org.junit.jupiter.api.Test;
+package nat; //same package as the class being tested
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class ToDoTest {
