--- src/main/java/Nat/Duke.java
@@ -1,14 +1,0 @@
-package Nat;
-
-public class Duke {
-    public static void main(String[] args) {
-        System.out.println("Hello!");
-    }
-
-    /**
-     * Generates a response for the user's chat message.
-     */
-    public String getResponse(String input) {
-        return "Nat.Duke heard: " + input;
-    }
-}
