+++ src/main/java/data/data.txt
@@ -1,5 +1,4 @@
 E | 0 | artbox | 4/1/6969 1420 | 31/12/9999 2359
 T | 0 | gym
 D | 1 | submission | 12/12/2001 1600
-T | 0 | homework
 T | 1 | watch lecture
