+++ src/main/java/nat/DialogBox.java
@@ -1,4 +1,4 @@
-package Nat;
+package nat;
 
 import java.io.IOException;
 import java.util.Collections;
