+++ src/main/java/nat/Event.java
@@ -1,4 +1,4 @@
-package Nat;
+package nat;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -10,17 +10,29 @@ public class Event extends Task {
     private LocalDateTime startDate;
     private LocalDateTime dueDate;
 
+    /**
+     * Constructor for Event class
+     * @param taskName Name of the task
+     * @param startDate Task start date
+     * @param dueDate Task deadline date
+     */
     public Event(String taskName, String startDate, String dueDate) {
         super(taskName);
         this.startDate = this.parseDateTime(startDate);
         this.dueDate = this.parseDateTime(dueDate);
     }
 
+    /**
+     * Transform the date input from String into LocalDateTime
+     */
     private LocalDateTime parseDateTime(String dateTime) {
         DateTimeFormatter inputFormat = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
         return LocalDateTime.parse(dateTime, inputFormat);
     }
 
+    /**
+     * Override and return the parent class Task save format
+     */
     @Override
     public String toSaveFormat() {
         // Format: "T | 1 | Read a book | 5pm | 9pm"
@@ -28,6 +40,9 @@ public class Event extends Task {
         return super.toSaveFormat() + " | " + this.startDate.format(format) + " | " + this.dueDate.format(format);
     }
 
+    /**
+     * Override and return the Event task type
+     */
     @Override
     public String getTaskType() {
         return "E";
