+++ src/main/java/nat/Launcher.java
@@ -1,4 +1,4 @@
-package Nat;
+package nat;
 
 import javafx.application.Application;
