+++ src/main/java/nat/Main.java
@@ -1,4 +1,4 @@
-package Nat;
+package nat;
 
 import java.io.IOException;
 
@@ -22,7 +22,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(nat);  // inject the Nat.Duke instance
+            fxmlLoader.<MainWindow>getController().setDuke(nat); // inject the Nat.Duke instance
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
