+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@
      _| |_\   |_  _/ /   \ \_  _| |_
     |_____|\____||____| |____||_____|
     ____________________________________________________________
-     Good day! I'm Nat!
+     Good day! I'm nat!
      How may I assist you today?
     ____________________________________________________________
      Here are the tasks in your list:
